pyproject.toml
[project]
name = "dmelhubert"
version = "0.1.0"
description = ""
authors = [{ name = "nicolvisser", email = "nicolvisser@yahoo.com" }]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "xformers (>=0.0.29.post3,<0.0.30)",
    "lightning (>=2.5.1,<3.0.0)",
    "wandb (>=0.19.9,<0.20.0)",
    "torchaudio (>=2.6.0,<3.0.0)",
    "soundfile (>=0.13.1,<0.14.0)",
    "zerospeech-libriabx2 (>=0.9.8,<0.10.0)",
    "faiss-cpu (>=1.10.0,<2.0.0)",
    "scikit-learn (>=1.6.1,<2.0.0)",
    "argparse-dataclass (>=2.0.0,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"
matplotlib = "^3.10.1"
